Решение задачи 9
